requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
exceptiongroup==1.0.0rc9
flake8==5.0.4
h11==0.14.0
idna==3.4
mccabe==0.7.0
outcome==1.2.0
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
PySocks==1.7.1
python-dotenv==0.21.0
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
